requirements.txt
Django==1.10.1
django-braces==1.9.0
django-filter==0.15.2
django-oauth-toolkit==0.10.0
djangorestframework==3.4.7
Markdown==2.6.7
MySQL-python==1.2.5
oauthlib==1.0.3
six==1.10.0
